Reorganize and Refactor some recent node replacement logic
