+++ src/components/shared/ReactFlow/FlowCanvas/TaskNode/TaskNode.tsx
@@ -134,6 +134,7 @@ function generateOutputHandles(outputSpecs: OutputSpec[]): ReactElement[] {
 
 const ComponentTaskNode = ({ data, selected }: NodeProps) => {
   const { taskStatusMap } = useComponentSpec();
+
   const [isComponentEditorOpen, setIsComponentEditorOpen] = useState(false);
   const [isTaskDetailsSheetOpen, setIsTaskDetailsSheetOpen] = useState(false);
   const taskId = useMemo(
