+++ src/components/shared/ReactFlow/FlowSidebar/components/ComponentItem.tsx
@@ -1,16 +1,12 @@
 import { File } from "lucide-react";
 import type { DragEvent } from "react";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useMemo } from "react";
 
 import { ComponentDetailsDialog } from "@/components/shared/Dialogs";
 import { SidebarMenuItem } from "@/components/ui/sidebar";
+import useComponentFromUrl from "@/hooks/useComponentFromUrl";
 import { cn } from "@/lib/utils";
 import { EMPTY_GRAPH_COMPONENT_SPEC } from "@/providers/ComponentSpecProvider";
-import {
-  fetchAndStoreComponent,
-  generateDigest,
-  parseComponentData,
-} from "@/services/componentService";
 import { type ComponentItemFromUrlProps } from "@/types/componentLibrary";
 import type {
   ComponentReference,
@@ -18,7 +14,6 @@ import type {
   TaskSpec,
 } from "@/utils/componentSpec";
 import { getComponentName } from "@/utils/getComponentName";
-import { getComponentByUrl } from "@/utils/localforge";
 
 interface ComponentMarkupProps {
   url: string;
@@ -122,93 +117,21 @@ const ComponentMarkup = ({
 const ComponentItemFromUrl = ({ url }: ComponentItemFromUrlProps) => {
   if (!url) return null;
 
-  const [isLoading, setIsLoading] = useState(true);
-  const [componentSpec, setComponentSpec] = useState<ComponentSpec | null>(
-    null,
-  );
-  const [componentText, setComponentText] = useState<string | null>(null);
-  const [componentDigest, setComponentDigest] = useState<string | null>(null);
-  const [error, setError] = useState<string | null>(null);
+  const { isLoading, error, componentRef } = useComponentFromUrl(url);
 
-  useEffect(() => {
-    const controller = new AbortController();
-    const signal = controller.signal;
-
-    const loadComponent = async () => {
-      if (!url) return;
-
-      try {
-        setIsLoading(true);
-        setError(null);
-
-        const storedComponent = await getComponentByUrl(url);
-
-        if (storedComponent) {
-          if (signal.aborted) return;
-          // Parse the component data
-          const text = storedComponent.data;
-          setComponentText(text);
-
-          try {
-            // Parse the component spec from the text
-            const parsedSpec = parseComponentData(text);
-            if (parsedSpec) {
-              setComponentSpec(parsedSpec);
-              const digest = await generateDigest(text);
-              setComponentDigest(digest);
-              setIsLoading(false);
-              return;
-            }
-          } catch (err) {
-            console.error("Error parsing component from local storage:", err);
-            // Fall through to network fetch
-          }
-        }
-
-        // If component doesn't exist in storage or parsing failed, fetch and store it
-        const spec = await fetchAndStoreComponent(url);
-        if (signal.aborted) return;
-
-        if (spec) {
-          setComponentSpec(spec);
-
-          // Get the stored component to get the text
-          const updatedComponent = await getComponentByUrl(url);
-          if (updatedComponent && !signal.aborted) {
-            const text = updatedComponent.data;
-            setComponentText(text);
-            const digest = await generateDigest(text);
-            setComponentDigest(digest);
-          }
-        } else {
-          throw new Error("Failed to load component specification");
-        }
-      } catch (err) {
-        if (!signal.aborted) {
-          setError(err instanceof Error ? err.message : String(err));
-        }
-      } finally {
-        if (!signal.aborted) {
-          setIsLoading(false);
-        }
-      }
-    };
-
-    loadComponent();
-    return () => controller.abort();
-  }, [url]);
+  const { spec, digest, text } = componentRef;
 
   const displayName = useMemo(
-    () => getComponentName({ spec: componentSpec ?? undefined, url }),
-    [componentSpec, url],
+    () => getComponentName({ spec: spec ?? undefined, url }),
+    [spec, url],
   );
 
   return (
     <ComponentMarkup
       url={url}
-      componentSpec={componentSpec || EMPTY_GRAPH_COMPONENT_SPEC}
-      componentDigest={componentDigest || ""}
-      componentText={componentText || ""}
+      componentSpec={spec || EMPTY_GRAPH_COMPONENT_SPEC}
+      componentDigest={digest || ""}
+      componentText={text || ""}
       displayName={displayName}
       isLoading={isLoading}
       error={error}
